Avoid Double instantiation
As collections in java may contain only Objects, primitive doubles can't be stored
directly. For each double an object is instantiated (of type Double) which is not efficient
if evaluate() method is called too often. Instead it's better to store a container object
with a writable field. If new value is set the field is update and no memory allocation
happens.